First attempt at class extraction.
Honestly, it would have been more sane to have just rewritten everything.
